# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 72
     */
  password: string;
}

export interface LoginInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 72
     */
  password: string;
}

export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  key: string;
  description: string;
  color: string;
  taskCount: number;
  completedTaskCount: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  key: string;
  description?: string;
  color?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  backlog: 'backlog',
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface Task {
  id: number;
  projectId: number;
  title: string;
  description: string;
  status: TaskStatus;
  priority: TaskPriority;
  assignee: string;
  /** @nullable */
  dueDate: string | null;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  backlog: 'backlog',
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskInput {
  projectId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  assignee?: string;
  /** @nullable */
  dueDate?: string | null;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  backlog: 'backlog',
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  assignee?: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface DashboardSummary {
  projectCount: number;
  openTaskCount: number;
  completedTaskCount: number;
  inProgressTaskCount: number;
  overdueTaskCount: number;
}

export type ActivityKind = typeof ActivityKind[keyof typeof ActivityKind];


export const ActivityKind = {
  task_created: 'task_created',
  task_completed: 'task_completed',
  project_created: 'project_created',
  task_updated: 'task_updated',
} as const;

export interface Activity {
  id: number;
  kind: ActivityKind;
  message: string;
  projectName: string;
  createdAt: string;
}

export type ListTasksParams = {
projectId?: number;
status?: ListTasksStatus;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  backlog: 'backlog',
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;