# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: auth
    description: Authentication and session operations
  - name: health
    description: Health operations
  - name: projects
    description: Project management operations
  - name: tasks
    description: Task management operations
  - name: dashboard
    description: Workspace overview operations
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start an authenticated session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Session ended
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current authenticated account
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: number
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [backlog, todo, in_progress, in_review, done]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get workspace summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /activity:
    get:
      operationId: listActivity
      tags: [dashboard]
      summary: List recent workspace activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: number
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: number
  schemas:
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 72
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
          maxLength: 72
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Project:
      type: object
      required: [id, name, key, description, color, taskCount, completedTaskCount]
      properties:
        id:
          type: number
        name:
          type: string
        key:
          type: string
        description:
          type: string
        color:
          type: string
        taskCount:
          type: number
        completedTaskCount:
          type: number
    ProjectInput:
      type: object
      required: [name, key]
      properties:
        name:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
    Task:
      type: object
      required: [id, projectId, title, description, status, priority, assignee, dueDate]
      properties:
        id:
          type: number
        projectId:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date
    TaskInput:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: number
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date
    DashboardSummary:
      type: object
      required: [projectCount, openTaskCount, completedTaskCount, inProgressTaskCount, overdueTaskCount]
      properties:
        projectCount:
          type: number
        openTaskCount:
          type: number
        completedTaskCount:
          type: number
        inProgressTaskCount:
          type: number
        overdueTaskCount:
          type: number
    Activity:
      type: object
      required: [id, kind, message, projectName, createdAt]
      properties:
        id:
          type: number
        kind:
          type: string
          enum: [task_created, task_completed, project_created, task_updated]
        message:
          type: string
        projectName:
          type: string
        createdAt:
          type: string
          format: date-time